backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy < 2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
scikit-image==0.25.2
scipy==1.15.3
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tifffile==2025.5.10
torch == 2.3.0
torchvision == 0.18.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
wsproto==1.2.0
